requirements.txt

django-debug-toolbar
django~=3.1.2


djangorestframework~=3.12.1